ShotImpactUpdate: send with greater precision
Use impactTPrecision for sending/receiving impact time
Change getImpactT() return type to double
client: Change types accordingly
client: GameApplet.create_shot_thread(): take the update instead of time

+++ src/re/neutrino/java_tanks/types/Impact.java
@@ -22,7 +22,7 @@ public class Impact {
 
 	public static Impact fromUpdate(
 			ShotImpactUpdate u, Player player, Shot shot, Config config) {
-		short time = u.getImpactT();
+		double time = u.getImpactT();
 		return new Impact(
 				Shot.getShotPos(player, shot, time, config).round(),
 				time);
